feat: Updated login mechanism

+++ src/pages/auth/LoginPage.tsx
@@ -1,9 +1,12 @@
 import React, {useEffect} from 'react';
-import {CircleStackIcon} from '@heroicons/react/24/outline';
 import {useGoogleAuth} from "../../hooks/firebase/useGoogleAuth.ts";
 import {useNavigate} from "react-router-dom";
 import appIcon from '../../assets/appicon.png'
 import {USER_KEY} from "../../services/StorageService.ts";
+// import { useFirestoreSaveCollection } from '../../hooks/firebase/useFirestoreSaveCollection.ts';
+import { db, FirestoreCollection } from '../../services/FirebaseService.ts';
+import { UserWater } from '../../models/UserWater.ts';
+import { doc, getDoc, setDoc, Timestamp } from 'firebase/firestore';
 
 const LoginWithGoogle: React.FC = () => {
   const navigate = useNavigate();
@@ -12,12 +15,47 @@ const LoginWithGoogle: React.FC = () => {
     login().then()
   };
 
+  // const {save} = useFirestoreSaveCollection<UserWater>(FirestoreCollection.users)
   const {user, login} = useGoogleAuth();
 
   useEffect(() => {
     if (!user) return
-    localStorage.setItem(USER_KEY, JSON.stringify(user))
-    navigate('/home')
+
+    const data: UserWater = {
+      uid: user.uid,
+      email: user.email ?? '',
+      displayName: user.displayName ?? '',
+      photoURL: user.photoURL ?? '',
+      createdAt: Timestamp.now()
+    }
+
+    const userRef = doc(db, FirestoreCollection.users, user.uid);
+    getDoc(userRef).then((docSnap) => {
+      if (docSnap.exists()) {
+        // Update only fields you want, do not overwrite createdAt
+        setDoc(userRef, {
+          uid: user.uid,
+          email: user.email ?? '',
+          displayName: user.displayName ?? '',
+          photoURL: user.photoURL ?? ''
+        }, { merge: true }).then(() => {
+          localStorage.setItem(USER_KEY, JSON.stringify(user));
+          navigate('/home');
+        });
+      } else {
+        // Create new user with createdAt
+        setDoc(userRef, {
+          uid: user.uid,
+          email: user.email ?? '',
+          displayName: user.displayName ?? '',
+          photoURL: user.photoURL ?? '',
+          createdAt: Timestamp.now()
+        }).then(() => {
+          localStorage.setItem(USER_KEY, JSON.stringify(user));
+          navigate('/home');
+        });
+      }
+    });
   }, [user]);
   return (
     <div className="flex items-center justify-center">
